feat: recruit search and readAvailableRecruit fix - 검색 api 개발 - 공고-기관의 1번 api 에러 발견 및 수정

+++ backend/ttabong/src/main/java/com/ttabong/controller/search/SearchController.java
@@ -16,8 +16,8 @@ public class SearchController {
 
     @GetMapping("/templates")
     public ResponseEntity<RecruitResponseDto> searchTemplates(
-            @RequestParam(required = false) Long cursor,
-            @RequestParam(defaultValue = "10") int limit,
+            @RequestParam(required = false) Integer cursor,
+            @RequestParam(defaultValue = "10") Integer limit,
             @RequestBody RecruitRequestDto request) {
 
         RecruitResponseDto responseDto = searchService.searchTemplates(request, cursor, limit);
