+++ backend/ttabong/src/main/java/com/ttabong/service/search/SearchService.java
@@ -4,5 +4,5 @@ import com.ttabong.dto.search.RecruitRequestDto;
 import com.ttabong.dto.search.RecruitResponseDto;
 
 public interface SearchService {
-    RecruitResponseDto searchTemplates(RecruitRequestDto request, Long cursor, int limit);
+    RecruitResponseDto searchTemplates(RecruitRequestDto request, Integer cursor, Integer limit);
 }
